testing/web-platform/mozilla/tests/service-workers/service-worker/getregistrations.sub.html

<!DOCTYPE html>
<title>Service Worker: getRegistrations()</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script src="resources/get-host-info.sub.js"></script>
<script src="../fetch/resources/fetch-test-helpers.sub.js"></script>
<script>
// Purge the existing registrations for the origin.
// getRegistrations() is used in order to avoid adding additional complexity
// e.g. adding an internal function.
sequential_promise_test(function(t) {
    return navigator.serviceWorker.getRegistrations()
      .then(function(registrations) {
          return registrations.reduce(function(sequence, registration) {
              return sequence.then(function() {
                  return registration.unregister();
                });
            }, Promise.resolve());
        });
  }, 'Purge the existing registrations.');

sequential_promise_test(function(t) {
    return navigator.serviceWorker.getRegistrations()
      .then(function(value) {
          assert_array_equals(
            value,
            [],
            'getRegistrations should resolve with an empty array.');
        });
  }, 'getRegistrations');

sequential_promise_test(function(t) {
    var scope = 'resources/scope/getregistrations/normal';
    var script = 'resources/empty-worker.js';
    var registrations = [];
    return service_worker_unregister_and_register(t, script, scope)
      .then(function(r) {
          registrations.push(r);
          return navigator.serviceWorker.getRegistrations();
        })
      .then(function(value) {
          assert_array_equals(
            value,
            registrations,
            'getRegistrations should resolve with array of registrations.');
          return service_worker_unregister(t, scope);
        });
  }, 'Register then getRegistrations');

sequential_promise_test(function(t) {
    var scope1 = 'resources/scope/getregistrations/scope1';
    var scope2 = 'resources/scope/getregistrations/scope2';
    var script = 'resources/empty-worker.js';
    var registrations = [];
    return service_worker_unregister_and_register(t, script, scope1)
      .then(function(r) {
          registrations.push(r);
          return service_worker_unregister_and_register(t, script, scope2);
        })
      .then(function(r) {
          registrations.push(r);
          return navigator.serviceWorker.getRegistrations();
        })
      .then(function(value) {
          assert_array_equals(
            value,
            registrations,
            'getRegistrations should resolve with array of registrations.');
          return service_worker_unregister(t, scope1);
        })
      .then(function() {
          return service_worker_unregister(t, scope2);
        });
  }, 'Register multiple times then getRegistrations');

sequential_promise_test(function(t) {
    var scope = 'resources/scope/getregistrations/register-unregister';
    var script = 'resources/empty-worker.js';
    return service_worker_unregister_and_register(t, script, scope)
      .then(function(registration) {
          return registration.unregister();
        })
      .then(function() {
          return navigator.serviceWorker.getRegistrations();
        })
      .then(function(value) {
          assert_array_equals(
            value,
            [],
            'getRegistrations should resolve with an empty array.');
        });
  }, 'Register then Unregister then getRegistrations');

sequential_promise_test(function(t) {
    // Top-level window's origin is http://{{domains[www]}}:{{ports[http][0]}}
    // Set frame's origin to http://{{host}}:{{ports[http][0]}}
    var host_info = get_host_info();
    var frame_url = host_info['HTTP_REMOTE_ORIGIN'] +
                    '/service-worker/resources/frame-for-getregistrations.sub.html';
    var scope = 'resources/scope-for-getregistrations';
    var script = 'resources/empty-worker.js';
    var frame;
    var registrations = [];

    return with_iframe(frame_url)
      .then(function(f) {
          // frame registered its registration scoped
          // http://{{host}}:{{ports[http][0]}}/service-worker/resources/scope-for-getregistrations
          frame = f;
          // Top-level window registers its registration scoped
          // http://{{domains[www]}}:{{ports[http][0]}}/service-worker/resources/scope-for-getregistrations
          return service_worker_unregister_and_register(t, script, scope);
        })
      .then(function(r) {
          registrations.push(r);
          return navigator.serviceWorker.getRegistrations();
        })
      .then(function(value) {
          assert_array_equals(
            value,
            registrations,
            'getRegistrations should only return same origin registrations.');

          var channel = new MessageChannel();
          var resolve;
          var p = new Promise(function(r) { resolve = r; });

          channel.port1.onmessage = function(e) {
            if (e.data == 'unregistered')
              resolve();
          };
          frame.contentWindow.postMessage('unregister', '*', [channel.port2]);
          return p;
        })
      .then(function() {
          frame.remove();
          return service_worker_unregister(t, scope);
        });
  }, 'getRegistrations promise resolves only with same origin registrations.');

sequential_promise_test_done();
done();
</script>
